+++ fpp_analysis/src/passes/state_machine/check_transition_graph/mod.rs
@@ -6,6 +6,19 @@ pub use check_choice_cycles::*;
 
 mod check_tg_reachability;
 pub use check_tg_reachability::*;
+use fpp_ast::DefStateMachine;
 
-mod check_transition_graph;
-pub use check_transition_graph::*;
+use crate::{analyzers::state_machine::SmResult, semantics::state_machine::StateMachineAnalysis};
+
+/// Compute and check the transition graph
+pub struct CheckTransitionGraph;
+
+impl CheckTransitionGraph {
+    pub fn def_state_machine(sma: StateMachineAnalysis, node: &DefStateMachine) -> SmResult {
+        let mut sma = ConstructTransitionGraph::def_state_machine(sma, node)?;
+        CheckTGReachability::state_machine_analysis(&sma)?;
+        CheckChoiceCycles::state_machine_analysis(&sma)?;
+        sma.reverse_transition_graph = sma.transition_graph.get_reverse_graph();
+        Ok(sma)
+    }
+}
